EDS[FIX]: Created Utils for convert currency to another and get his amount in globalTransaction.

+++ data/src/main/java/com/example/data/model/mapper/ModelsMapper.kt
@@ -21,7 +21,7 @@ fun transactionResponseToTransactionLocal(transactionResponseList: ArrayList<Tra
         transactionLocalList.add(
             TransactionLocal(
                 sku = transaction.sku,
-                amount = transaction.amount.toFloat(),
+                amount = transaction.amount,
                 currency = transaction.currency
             )
         )
